requirements.txt
amqp==5.2.0
apimatic-core==0.2.3
apimatic-core-interfaces==0.1.3
apimatic-requests-client-adapter==0.1.4
asgiref==3.6.0
async-timeout==4.0.3
billiard==4.2.0
CacheControl==0.12.11
celery==5.3.6
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.0
deprecation==2.1.0
Django==4.2.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.0.0
django-timezone-field==6.1.0
djangorestframework==3.14.0
enum34==1.1.10
idna==3.4
jsonpickle==3.0.1
jsonpointer==2.3
kombu==5.3.4
msgpack==1.0.5
packaging==23.1
prompt-toolkit==3.0.41
PyJWT==2.7.0
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2023.3
redis==5.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
squareup==27.0.0.20230517
tzdata==2023.3
urllib3==2.0.2
vine==5.1.0
wcwidth==0.2.12
